# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Backend",
        "version": "1.0.0",
        "description": "Sistema completo de adoção, ONGs, eventos, necessidades e doações."
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
...
    "paths": {
        "/inserir": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar usuário",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsuarioInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/atualizar": {
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar usuário",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Usuario"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            }
        },
        "/deletar": {
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deletar usuário",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deletado"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Login"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado"
                    }
                }
            }
        },
        "/inserirAnimal": {
            "post": {
                "tags": [
                    "Animais"
                ],
                "summary": "Cadastrar animal",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnimalInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/atualizarAnimal": {
            "put": {
                "tags": [
                    "Animais"
                ],
                "summary": "Atualizar animal",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Animal"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            }
        },
        "/deletarAnimal": {
            "delete": {
                "tags": [
                    "Animais"
                ],
                "summary": "Deletar animal",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_animal": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deletado"
                    }
                }
            }
        },
        "/inserirOng": {
            "post": {
                "tags": [
                    "ONGs"
                ],
                "summary": "Criar ONG",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OngInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/atualizarOng": {
            "put": {
                "tags": [
                    "ONGs"
                ],
                "summary": "Atualizar ONG",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ong"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            }
        },
        "/deletarOng": {
            "delete": {
                "tags": [
                    "ONGs"
                ],
                "summary": "Deletar ONG",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_ong": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deletado"
                    }
                }
            }
        },
        "/inserirDica": {
            "post": {
                "tags": [
                    "Dicas"
                ],
                "summary": "Criar dica",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DicaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/atualizarDica": {
            "put": {
                "tags": [
                    "Dicas"
                ],
                "summary": "Atualizar dica",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Dica"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            }
        },
        "/inserirDoacao": {
            "post": {
                "tags": [
                    "Doações"
                ],
                "summary": "Criar doação",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DoacaoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/atualizarDoacao": {
            "put": {
                "tags": [
                    "Doações"
                ],
                "summary": "Atualizar doação",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Doacao"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado"
                    }
                }
            }
        },
        "/deletarDoacao": {
            "delete": {
                "tags": [
                    "Doações"
                ],
                "summary": "Deletar doação",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id_doacao": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deletado"
                    }
                }
            }
        },
        "/inserirNecessidade": {
            "post": {
                "tags": [
                    "Necessidades"
                ],
                "summary": "Criar necessidade",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NecessidadeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Login": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "senha": {
                        "type": "string"
                    }
                }
            },
            "Usuario": {
                "type": "object",
                "properties": {
                    "id_usuario": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "senha": {
                        "type": "string"
                    }
                }
            },
            "UsuarioInput": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "senha": {
                        "type": "string"
                    }
                }
            },
...
            "AnimalInput": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "especie": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}